+++ src/Stub/Answer/AnswerInterface.php
@@ -17,25 +17,11 @@ namespace Eloquent\Phony\Stub\Answer;
 interface AnswerInterface
 {
     /**
-     * Set the primary request.
-     *
-     * @param CallRequestInterface $primaryRequest The primary request.
-     */
-    public function setPrimaryRequest(CallRequestInterface $primaryRequest);
-
-    /**
      * Get the primary request.
      *
      * @return CallRequestInterface|null The primary request, or null if none has been set.
      */
     public function primaryRequest();
-
-    /**
-     * Add a secondary request.
-     *
-     * @param CallRequestInterface $secondaryRequest The secondary request.
-     */
-    public function addSecondaryRequest(CallRequestInterface $secondaryRequest);
 
     /**
      * Get the secondary requests.