+++ test/suite/Mock/Proxy/Factory/ProxyFactoryTest.php
@@ -14,7 +14,6 @@ namespace Eloquent\Phony\Mock\Proxy\Factory;
 use Eloquent\Phony\Assertion\Recorder\AssertionRecorder;
 use Eloquent\Phony\Assertion\Renderer\AssertionRenderer;
 use Eloquent\Phony\Invocation\Invoker;
-use Eloquent\Phony\Matcher\WildcardMatcher;
 use Eloquent\Phony\Mock\Builder\MockBuilder;
 use Eloquent\Phony\Mock\Proxy\Stubbing\StaticStubbingProxy;
 use Eloquent\Phony\Mock\Proxy\Stubbing\StubbingProxy;
@@ -32,14 +31,12 @@ class ProxyFactoryTest extends PHPUnit_Framework_TestCase
         $this->stubVerifierFactory = new StubVerifierFactory();
         $this->assertionRenderer = new AssertionRenderer();
         $this->assertionRecorder = new AssertionRecorder();
-        $this->wildcardMatcher = new WildcardMatcher();
         $this->invoker = new Invoker();
         $this->subject = new ProxyFactory(
             $this->stubFactory,
             $this->stubVerifierFactory,
             $this->assertionRenderer,
             $this->assertionRecorder,
-            $this->wildcardMatcher,
             $this->invoker
         );
     }
@@ -50,7 +47,6 @@ class ProxyFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertSame($this->stubVerifierFactory, $this->subject->stubVerifierFactory());
         $this->assertSame($this->assertionRenderer, $this->subject->assertionRenderer());
         $this->assertSame($this->assertionRecorder, $this->subject->assertionRecorder());
-        $this->assertSame($this->wildcardMatcher, $this->subject->wildcardMatcher());
         $this->assertSame($this->invoker, $this->subject->invoker());
     }
 
@@ -62,7 +58,6 @@ class ProxyFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertSame(StubVerifierFactory::instance(), $this->subject->stubVerifierFactory());
         $this->assertSame(AssertionRenderer::instance(), $this->subject->assertionRenderer());
         $this->assertSame(AssertionRecorder::instance(), $this->subject->assertionRecorder());
-        $this->assertSame(WildcardMatcher::instance(), $this->subject->wildcardMatcher());
         $this->assertSame(Invoker::instance(), $this->subject->invoker());
     }
 
@@ -84,7 +79,6 @@ class ProxyFactoryTest extends PHPUnit_Framework_TestCase
             $this->stubVerifierFactory,
             $this->assertionRenderer,
             $this->assertionRecorder,
-            $this->wildcardMatcher,
             $this->invoker
         );
         $actual = $this->subject->createStubbing($mock, 'label');
@@ -177,7 +171,6 @@ class ProxyFactoryTest extends PHPUnit_Framework_TestCase
             $this->stubVerifierFactory,
             $this->assertionRenderer,
             $this->assertionRecorder,
-            $this->wildcardMatcher,
             $this->invoker
         );
         $actual = $this->subject->createStubbingStatic($class);