+++ test/suite/Mock/Proxy/Stubbing/StubbingProxyTest.php
@@ -16,7 +16,6 @@ use Eloquent\Phony\Assertion\Renderer\AssertionRenderer;
 use Eloquent\Phony\Event\EventCollection;
 use Eloquent\Phony\Feature\FeatureDetector;
 use Eloquent\Phony\Invocation\Invoker;
-use Eloquent\Phony\Matcher\WildcardMatcher;
 use Eloquent\Phony\Mock\Builder\MockBuilder;
 use Eloquent\Phony\Stub\Factory\StubFactory;
 use Eloquent\Phony\Stub\Factory\StubVerifierFactory;
@@ -31,7 +30,6 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
         $this->stubVerifierFactory = new StubVerifierFactory();
         $this->assertionRenderer = new AssertionRenderer();
         $this->assertionRecorder = new AssertionRecorder();
-        $this->wildcardMatcher = new WildcardMatcher();
         $this->invoker = new Invoker();
 
         $this->featureDetector = FeatureDetector::instance();
@@ -50,7 +48,6 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
             $this->stubVerifierFactory,
             $this->assertionRenderer,
             $this->assertionRecorder,
-            $this->wildcardMatcher,
             $this->invoker
         );
 
@@ -72,11 +69,11 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
         $this->assertSame($this->state->stubs, $this->subject->stubs());
         $this->assertSame($this->state->isFull, $this->subject->isFull());
         $this->assertSame($this->state->label, $this->subject->label());
+        $this->assertTrue($this->subject->isAdaptable());
         $this->assertSame($this->stubFactory, $this->subject->stubFactory());
         $this->assertSame($this->stubVerifierFactory, $this->subject->stubVerifierFactory());
         $this->assertSame($this->assertionRenderer, $this->subject->assertionRenderer());
         $this->assertSame($this->assertionRecorder, $this->subject->assertionRecorder());
-        $this->assertSame($this->wildcardMatcher, $this->subject->wildcardMatcher());
         $this->assertSame($this->invoker, $this->subject->invoker());
     }
 
@@ -93,7 +90,6 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
         $this->assertSame(StubVerifierFactory::instance(), $this->subject->stubVerifierFactory());
         $this->assertSame(AssertionRenderer::instance(), $this->subject->assertionRenderer());
         $this->assertSame(AssertionRecorder::instance(), $this->subject->assertionRecorder());
-        $this->assertSame(WildcardMatcher::instance(), $this->subject->wildcardMatcher());
         $this->assertSame(Invoker::instance(), $this->subject->invoker());
     }
 
@@ -103,10 +99,17 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
 
         $this->assertSame($this->subject, $this->subject->setLabel(null));
         $this->assertNull($this->subject->label());
-
-        $this->subject->setLabel($this->state->label);
-
+        $this->assertSame($this->subject, $this->subject->setLabel($this->state->label));
         $this->assertSame($this->state->label, $this->subject->label());
+    }
+
+    public function testSetIsAdaptable()
+    {
+        $this->setUpWith('Eloquent\Phony\Test\TestClassA');
+
+        $this->assertTrue($this->subject->isAdaptable());
+        $this->assertSame($this->subject, $this->subject->setIsAdaptable(false));
+        $this->assertFalse($this->subject->isAdaptable());
     }
 
     public function testFull()
@@ -171,7 +174,7 @@ class StubbingProxyTest extends PHPUnit_Framework_TestCase
     {
         $this->setUpWith('Eloquent\Phony\Test\TestClassA');
 
-        $this->setExpectedException('Eloquent\Phony\Mock\Proxy\Exception\UndefinedPropertyException');
+        $this->setExpectedException('Eloquent\Phony\Mock\Exception\UndefinedMethodStubException');
         $this->subject->nonexistent;
     }
 
@@ -234,17 +237,19 @@ EOD;
     {
         $this->setUpWith('Eloquent\Phony\Test\TestClassA');
         $actual = $this->subject->testClassAMethodA();
+        $actual->returns();
 
         $this->assertInstanceOf('Eloquent\Phony\Stub\StubVerifier', $actual);
         $this->assertSame($actual, $this->subject->testClassAMethodA);
         $this->assertSame($actual, $this->subject->state()->stubs->testclassamethoda);
+        $this->assertSame($actual, $this->subject->testClassAMethodA()->returns());
     }
 
     public function testMagicCallFailure()
     {
         $this->setUpWith('Eloquent\Phony\Test\TestClassA');
 
-        $this->setExpectedException('Eloquent\Phony\Mock\Proxy\Exception\UndefinedMethodException');
+        $this->setExpectedException('Eloquent\Phony\Mock\Exception\UndefinedMethodStubException');
         $this->subject->nonexistent();
     }