+++ CHANGELOG.md
@@ -18,11 +18,15 @@
 - **[BC BREAK]** Removed the `$self` and `$defaultAnswerCallback` arguments from
   `stub()` ([#123]).
 - **[IMPROVED]** Made API setter style methods fluent ([#98]).
+- **[IMPROVED]** Instance handles are automatically adapted when stubbing and
+  verifying ([#126]).
+- **[IMPROVED]** Added checks for unused stub criteria ([#126]).
 - **[FIXED]** Fixed bug when mocking traits with magic call methods ([#127]).
 
 [#98]: https://github.com/eloquent/phony/issues/98
 [#117]: https://github.com/eloquent/phony/issues/117
 [#123]: https://github.com/eloquent/phony/issues/123
+[#126]: https://github.com/eloquent/phony/issues/126
 [#127]: https://github.com/eloquent/phony/issues/127
 
 ## 0.7.0 (2015-12-20)